Per the skill: mini commit message before moving to the test discovery step.
  fix(test): rename Test.MassSpectrometry.Deconvolution to avoid prod-namespace shadowing

  Two test files in Test/MassSpectrometryTests/Deconvolution/ declared
  namespace Test.MassSpectrometry.Deconvolution, which shadowed the
  production-library MassSpectrometry namespace from inside the test
  project. C#'s relative-namespace resolution then mis-resolved
  "MassSpectrometry.DissociationType" (and similar references) to
  Test.MassSpectrometry.DissociationType -- a type that doesn't exist --
  causing 18 cascading CS0234 errors in unrelated files (TestDatabaseLoaders,
  BackboneFragmentation, FragmentModificationTests, etc.).

  Rename to Test.MassSpectrometryTests.Deconvolution to match the surrounding
  files in the same directory and the directory name itself. Update the four
  "using static" references that pointed at the old namespace.

  Files changed:
  - DeconvolutionTestHelpers.cs (namespace)
  - TestDeconvolutionScorerSpectrumAware.cs (namespace + using static)
  - TestDeconvolutionScorerNegativeCharge.cs (using static)
  - TestDeconvolutionScorerUnit.cs (using static)
  - TestIsotopicEnvelopeExtensions.cs (using static)
